Moved implementation classes related to daemon execution to separated package
